move messages to shared
